requirements.txt
arcgis==1.8.2
argon2-cffi @ file:///C:/ci/argon2-cffi_1596826740294/work
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
bmi-arcgis-restapi==1.0.5
boostrtrees @ file:///C:/Users/Larkin/planning_llh_bgc/pyboostrtree/dist/boostrtrees-0.0.1a1.tar.gz
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
Cython @ file:///C:/ci/cython_1594834059511/work
decorator==4.4.2
defusedxml==0.6.0
dill @ file:///tmp/build/80754af9/dill_1593202311028/work
entrypoints==0.3
graphviz==0.14
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593431910144/work
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi @ file:///C:/ci/jedi_1596472773465/work
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==21.2.0
kiwisolver==1.2.0
lerc @ file:///Volumes/JetDrive/virtualenv/miniconda/conda-bld/lerc_1594078116229/work
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1592846129657/work
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
notebook @ file:///C:/ci/notebook_1596840158498/work
ntlm-auth==1.4.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1596233948451/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.4
pandas @ file:///C:/ci/pandas_1596825959733/work
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1594299032796/work
plotly @ file:///tmp/build/80754af9/plotly_1597716491413/work
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyproj==2.2.0
pyro-ppl==0.4.1
Pyro4 @ file:///home/conda/feedstock_root/build_artifacts/pyro4_1588174595613/work
pyrsistent==0.16.0
pyshp==2.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
Rtree==0.9.4
scipy @ file:///C:/ci/scipy_1592930618155/work
selenium==3.141.0
Send2Trash==1.5.0
serpent==1.30.2
Shapely==1.7.1
six==1.15.0
soupsieve==2.0.1
terminado==0.8.3
testpath==0.4.4
torch==1.5.0
torchvision==0.6.0
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1596810128862/work
traitlets==4.3.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
winkerberos==0.7.0
zipp==3.1.0
